Redirect users to complete consent form and enrollment questionnaire
Redirect users to REDCap if their eligibility screening, consent form,
or enrollment questionnaire is not complete. Fortunately, we can use
REDCap's survey queue logic here and consistently point participants to
the eligibilty screening (first questionnaire in the queue). If that
survey is already complete, REDCap will redirect them to the next survey
in the queue.