Show users simple message if REDCap fails to fetch their data
Right now, the only exception we throw is for an AssertionError in the
(hopefully rare) case REDCap returns multiple matches for a given NetID,
which should be a secondary unique identifier in REDCap.

Catch all possible errors when fetching REDCap data, log the underlying
error message, and show the user a simple message to point them to
getting help from the support team.