Fix: Account for nonexistent instruments
In REDCap, instruments that don't exist will still have a value of an
empty string for the given instrument's `f"{instrument}_complete"` key.
Account for this by filtering out empty string values for the instrument
of interest, as the id3c `is_complete()` function is smart and will
return None if the instrument does not exist.