Raise custon InvalidNetId error
Instead of a base BadRequest, raise an InvalidNetId error when we want
to display feedback for incorrectly entered NetIds. Now we won't capture
other BadRequests Flask might raise or run into an issue down the road
if we want to raise BadRequests for different reasons.

Note that this still redacts NetID (PII) from the logs.