Redirect to employee view of KR form
Instead of generating the PT-facing survey link for the kiosk
registration form, drop the Lead Dawgs kiosk staff user inside of of the
REDCap employee-side of the Kiosk Registration form. This view has more
information about the current record which will better allow the kiosk
staff to verify that they're in the right place.