Collect request metrics for scraping by Prometheus
This will let us understand both the volume of requests and request rate
as well as how long requests to the various endpoints are taking, so we
can look at bottlenecks.

Uses a collector that supports multiple processes since in production we
use a hybrid multiprocess and threaded deployment with uWSGI.  This
requires a directory for collecting the metrics from each individual
process, which are then aggregated and exported.  Metrics are exported
at /metrics.

If no collection directory is defined, then metrics are not collected
and exported.

Disables the default "flask_" prefix on metric names in order to use the
standard Prometheus http_* metrics names as exported by other services.

Disables the histogram metric for request durations to result in a
summary metric instead, which is simpler to deal with on the query side
and should be enough for our needs.

Excludes /static/… paths to trim down the time series we store in
Prometheus.  We don't really care about metrics for static assets and
presume that they will Just Work.