Collect metrics specifically for requests we make to REDCap
The same sort of summary metric (total count and duration) being
captured for requests made to Musher, but for Musher's requests to
REDCap as identified by the containing function name.  I'm not sure that
function name is the best label here long-term, but I think it'll do for
now.

These will let us disentangle Musher itself from REDCap if we need/want
to.