Cache our most expensive REDCap lookup
The extremely slow (10-20s) fetch_participant() REDCap lookups during
high-volume periods (e.g. daily check-ins or kiosk surges) are creating
a huge bottleneck and leading to request queues filling up.  This is a
severe problem for the study.

The result of fetch_participant() is only cached if enrollment is
complete, which implies that the bare-bones information in the result is
very unlikely to change.  The majority of our volume is from enrolled
participants.

The meaning of the "fetch_participants" metric is preserved so we can
still observe accurate REDCap request timings, but it is expected to
happen much less often with caching.  A new "fetch_participants
(cached)" metric is exported for the whole (possibly-cached) lookup.