Add new test case for multiple tests in same week
Add a new unit test to make sure we're pulling up the correct target
KR instance for a PT who had a mail test kit ordered within the same
week. The motivation for adding this test case comes from reports of a
possible bug in the Lead Dawgs logic, but the test case passes.