Add (disabled) proof of concept for structured logging
Add a currently disabled structured logging handler. This writes out the logs
in json format, with a standardized set of metadata fields. These can be more
easily parsed and searched than unstructured string logs. The new handler writes
logs to a file, which will eventually be ingested into a centralized log store
along with logs from other applications.

The structured logs are a bit harder to read, so leave the existing loghandler as is,
with the idea of eventually adding in this structured log handler.