Require timestamp when linking to incomplete KR instance
Originally, musher was designed to link into incomplete KR forms if
they existed, to handle the case of someone starting to register at the
kiosk, and partially filled out without being saved as complete. This
way, we would link the kiosk staff back into that incomplete kiosk
registration so it could be completed.

However, this doesn't account for certain behavior in RECCap, namely,
that incomplete KR forms are getting created as a byproduct of a repeat
instance getting created with any other instruments. It appears that
when REDCap creates a repeat instance for another instrument, for
example, a daily attestation, any forms in that repeat instance event
with calculated variables get created and those calculated variables
get values set.

These REDCap created forms that exist but aren't really partially
complete from a human user will still have kiosk_registration_4c7f='0'
instead of ''. To account for these, and ignore these when looking for
the correct instance to link to, let's add the concept of a required
field for the encounter to be considered.

For kiosk registrations, we'll use the timestamp field nasal_swab_q,
and enforce that to exist for us to consider linking into an existing
KR as partially complete. This way we can ignore the REDCap created
forms with just calculated variables.