Drop unwanted IDP attributes keeping only netid
The study team has requested that we drop all attributes other than netid that we
receive from UW IDP. Those include the user's name, email, affilitation, and UW school
name.

This is being done so those fields will not autopopulate, forcing the user to manually
enter the information they are willing to share.