Add timeout setting for requests to REDCap server
Currently there is no timeout in place for these requests, so if the
REDCap server is having issues the requests can hang indefinitely and
cause number of available workers to drop to zero. By setting a timeout
the user will be redirected to the error page `something_went_wrong.html`.