Check content-type of request responses
Adding checks to confirm response from REDCap API are in JSON format. This is
to handle cases where REDCap may return a 200 status code with an HTML
response which contains an error message. In those situations the user will
now be routed to the general `something_went_wrong.html` page.