Set up git actions for doctest and unittest
Set up github actions for husky musher that will run
doctest and unittest on a branch push or pull request.
This file is based on the original version of the CI
pipeline that Thomas built for id3c and id3c-customizations.
This change also updates documentation for running tests so
developers run them correctly.