Lazy load RedCap project
Prior to this change, the RedCap project was initialized and
loaded at the top level of `redcap.py`, which had the side
effect of making network requests and accessing RedCap even
when this module was being run by the test suite. This meant
that we would have needed to store secrets in the musher
repository so that the RedCap API could be queried, just to run
tests which didn't even require API calls in the first place.

This change allows the RedCap project to be loaded lazily by
redcap utility functions when it is needed. To do this, I created
a project container, which is initialized to have its field
`redcap_project = None`. This way the RedCap project is no longer
initialized as soon as this file is imported by the testsuite. The
testsuite then sets the RedCap project to be a dummy project,
complete with the fields `redcap.py` expects a RedCap project to have.
When this file is not initialized by the test suite, the `redcap_project`
field in container will remain unitialized. When functions in this file
attempt to load the project for the first time, the project will first
check to make sure it has been loaded, and if not, load the RedCap
project as it would have before.

This logic flow allows us to avoid unnecessary calls to the RedCap
API during testing, while still preserving the loading of the
correct RedCap project in both the production and development
Flask environment.