Lazy load cache to prevent doctest failure
Since doctest was looking for wrapped classes inside the fanout cache,
which did not allow access of the `__wrapped__` attribute, doctests
were failing. Loading the fanout cache only when importing this file
from a flask context enables doctests to run without attempting to
access that invalid attribute from the fanout cache. This allows our
doctests to run without issue.

This change also includes some refactoring of previous changes to the
lazy loader wrapper class and associated comments to make it more clear
that it was loading both the cache and a redcap project.