Lazy load inside bin and add explicit import pathing
Updates cache scripts in `/bin` to use the lazy loaded cache and project
and updates their import structure to explicitly add the `husky_musher`
directory. Python does not pick up this directory by default, so in order
to use it it had to be added explicitly to the search path.