Disable LeadDawgs Route
HCT Y3 will no longer be using kiosk tests, so LeadDawgs is no longer needed. This change
adds a decorator to disable a route, rendering a 404 error for the route instead. If we
ever require LeadDawgs again, simply removing the `@route_disabled` decorator will allow
the route to be accessed again.