Add Script to Remove Deleted REDCap Records from Cache
When REDCap records are deleted, their associated entries must also be deleted
from the musher cache to ensure participants get directed to the correct
surveys. This was previously done manually, but if we parse the REDCap logs
to look for deleted records it is possible to do this as a nightly cleanup job
and not have to bother with manual removal.

Since REDCap logs don't have the netid (PHI!), we only have the record id of
the deleted record and have to find the netid to delete the cache record
somewhat inneficiently. I think this should be fine given when this job runs and
the current amount of cached netids. Deleted record ids are logged out to the
default musher log location at the info level.