Add Functions to Get REDCap Logs and Validate Responses to REDCap Utils
This change adds two functions to the REDCap utilities package. The first is gets
logging data for deleted records based on a start and end time. This will be used
to do a nightly cleanup of the Musher cache. The second is a wrapper for REDCap
post requests, which can be notoriously finicky. Sometimes the API will return
a 200 code but not the correct response value. We don't want our cache cleanup to
fail silently, so we check to make sure the response we expect is within the
actual response before using it.